% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renyi.R
\name{generalized_renyi_transform}
\alias{generalized_renyi_transform}
\title{Generalized Renyi Transform}
\usage{
generalized_renyi_transform(x, eta = NULL, zeta = NULL)
}
\arguments{
\item{x}{a vector of independent exponential random variables of the form \eqn{X_j = \eta_j Y_j + \zeta_j} where each \eqn{X_j} is an independent exponential random variable with rate 1}

\item{eta}{vector of scale parameters implicit in the construction of \code{x}: \code{eta[j]} = \eqn{\eta_j}}

\item{zeta}{vector of shift parameters implicit in the construction of \code{x}: \code{zeta[j]} = \eqn{\zeta_j}}
}
\value{
a list containing two elements
\describe{
  \item{`exps`}{a vector of independent standard exponentials where \code{exps[1]} is the exponential jump corresponding to \code{min(x)} and \code{tail(exps,1)} is the exponential jump corresponding to \code{max(x)}.}
  \item{`order`}{\code{order(x)}.}
}
}
\description{
A Generalization of Aldous Renyi's representation of exponential order statistics
}
\details{
Maps a vector of shifted and scaled independent exponential random variables to a sequence of standard independent exponential random variables based on the gaps (jumps) between the initial random variables
}
\examples{
# example code

a <- rchisq(10,1)
b <- rnorm(10)
xx <- a*rexp(10)+b
generalized_renyi_transform(xx, a, b)
}
\references{
Christ, R., Hall, I. and Steinsaltz, D.  (2024) "The Renyi Outlier Test", \href{https://arxiv.org/abs/2411.13542}{arXiv:2411.13542} . Available at: \doi{10.48550/arXiv.2411.13542}.
}
